Refactored application controller to includd concern for source feature
