+++ app/controllers/application_controller.rb
@@ -1,13 +1,5 @@
 class ApplicationController < ActionController::Base
   protect_from_forgery with: :exception
-  
-
-before_action :set_source
-
-   def set_source 
-      session[:source] = params[:q] if params[:q]
-   end 
+  include DeviseWhitelist
+  include SetSource
 end
-
-
-
