--- app/controllers/concerns/Devise_whitelist.rb
@@ -1,11 +1,0 @@
-module DeviseWhitelist
-  extend Activesupport::concern
- included do 
-   before_filter :configure_permitted_parameters, if: :devise_controller?
- end
-  
-  def configure_permitted_parameters
-  devise_parameter_sanitizer.permit(:sign_up, keys: [:name])
-  devise_parameter_sanitizer.permit(:account_update,keys: [:name])
-  end
-end
