+++ app/controllers/pages_controller.rb
@@ -1,7 +1,7 @@
 class PagesController < ApplicationController
   def home
     @posts = Blog.all
-    @Skills = Skill.all
+    @kills = Skill.all
   end
 
   def about
